+++ QGL_Graphics/qgl_depth_stencil.cpp
@@ -9,7 +9,7 @@ namespace qgl::graphics::gpu::render
       const content::buffers::DEPTH_STENCIL_BUFFER* buffer,
       graphics::graphics_device* dev_p,
       const graphics::window* wnd,
-      size_t frameIndex,
+      UINT frameIndex,
       const graphics::gpu::dsv_descriptor_heap* dsvHeap) :
       m_buffer(*buffer),
       m_width(wnd->width()),
